refactor: move collision and fill utilities to shared layer
This fixes a critical architectural violation where the core layer
(src/core/store/layout.ts) was importing from the features layer
(src/features/grid-editor/utils/). Features should depend on core,
not the other way around.

Changes:
- Move collision.ts from features/grid-editor/utils to shared/utils
- Move fill.ts from features/grid-editor/utils to shared/utils
- Update all imports across codebase to use new paths
- Remove now-empty grid-editor/utils/index.ts barrel file

The collision and fill utilities are cross-cutting concerns used by:
- Core store (layout.ts)
- Shared validation (validation.ts)
- Multiple features (bin-inspector, layers, grid-editor)

Moving them to shared/utils properly establishes them as reusable
utilities that can be imported by any layer without violating
the dependency graph.
