+++ src/core/store/layout.ts
@@ -3,8 +3,8 @@ import { immer } from 'zustand/middleware/immer';
 import type { Layout, Bin, Layer, Category, Drawer } from '@/core/types';
 import { createDefaultLayout, generateId, STAGING_ID, CONSTRAINTS, calcMaxGridUnits } from '@/core/constants';
 import { canPlaceBin, clamp } from '@/shared/utils/validation';
-import { fillAllWithSize, fillGaps } from '@/features/grid-editor/utils/fill';
-import { checkLayerReorderCollisions } from '@/features/grid-editor/utils/collision';
+import { fillAllWithSize, fillGaps } from '@/shared/utils/fill';
+import { checkLayerReorderCollisions } from '@/shared/utils/collision';
 import { useSettingsStore } from './settings';
 import { mlTracking } from '@/shared/analytics/useMLTracking';
 import type { Result, LayoutError, ValidationError } from '@/core/result';
