--- src/features/grid-editor/utils/index.ts
@@ -1,18 +1,0 @@
-// Grid editor utilities - barrel export
-export {
-  getDisplayLayers,
-  getLayerZStart,
-  getBin3DRect,
-  footprintsOverlap,
-  verticalRangesOverlap,
-  binsCollide,
-  getBlockedZones,
-  isInBlockedZone,
-  checkLayerReorderCollisions,
-} from './collision';
-
-export {
-  fillAllWithSize,
-  fillGaps,
-  getLayerCoverage,
-} from './fill';
