+++ src/shared/utils/collision.ts
