+++ src/test/fill.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { fillAllWithSize, fillGaps, getLayerCoverage } from '@/features/grid-editor/utils/fill';
+import { fillAllWithSize, fillGaps, getLayerCoverage } from '@/shared/utils/fill';
 import type { Layout } from '@/core/types';
 
 const createTestLayout = (): Layout => ({
